Move attribute converter to core instead of desc
