+++ meshroom/core/desc/__init__.py
@@ -29,7 +29,7 @@ from .shapeAttribute import (
 from .anySet import (
     AnySet
 )
-from .attributeConverter import (
+from ..attributeConverter import (
     AttributeConverter
 )
 from .computation import (
