+++ meshroom/core/plugins.py
@@ -18,6 +18,7 @@ from meshroom.core import desc
 from meshroom.core.desc.attribute import ValueTypeErrors
 from meshroom import _MESHROOM_ROOT
 from meshroom.core.desc.node import _MESHROOM_COMPUTE_DEPS
+from meshroom.core.attributeConverter import AttributeConverter
 
 
 def validateNodeDesc(nodeDesc: desc.BaseNode) -> list[tuple[str, ValueTypeErrors]]:
@@ -489,23 +490,23 @@ class AttributeConverterRegistry:
             f"Add converter class: {converterClass.__name__} "
             f"({converterClass.srcType.__name__} -> {converterClass.dstType.__name__})"
         )
-        if not issubclass(converterClass, desc.AttributeConverter):
+        if not issubclass(converterClass, AttributeConverter):
             raise TypeError(f"{converterClass} must subclass AttributeConverter")
         cls._converters[(converterClass.srcType, converterClass.dstType)].append(converterClass)
 
     @classmethod
-    def getAllConverters(cls) -> list[desc.AttributeConverter]:
+    def getAllConverters(cls) -> list[AttributeConverter]:
         return list(chain.from_iterable(cls._converters.values()))
 
     @classmethod
-    def getConverters(cls, srcType, dstType) -> list[desc.AttributeConverter]:
+    def getConverters(cls, srcType, dstType) -> list[AttributeConverter]:
         """ Get priority-ordered converters.
         """
         converters = cls._converters.get((srcType, dstType))
         return sorted(converters, key=lambda c: -c.priority)
 
     @classmethod
-    def getConverter(cls, srcType, dstType) -> desc.AttributeConverter:
+    def getConverter(cls, srcType, dstType) -> AttributeConverter:
         """ Get highest priority converter. """
         converters = cls.getConverters(srcType, dstType)
         if not converters:
