+++ meshroom/nodes/converters/BoolToInt.py
@@ -1,7 +1,8 @@
 from meshroom.core import desc
+from meshroom.core.attributeConverter import AttributeConverter
 
 
-class IntToBool(desc.AttributeConverter):
+class IntToBool(AttributeConverter):
     srcType = desc.BoolParam
     dstType = desc.IntParam
