+++ meshroom/nodes/converters/ChoiceToString.py
@@ -1,7 +1,8 @@
 from meshroom.core import desc
+from meshroom.core.attributeConverter import AttributeConverter
 
 
-class ChoiceToString(desc.AttributeConverter):
+class ChoiceToString(AttributeConverter):
     srcType = desc.ChoiceParam
     dstType = desc.StringParam
